+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailActions.kt
@@ -49,6 +49,9 @@ sealed class RoomDetailActions {
 
     data class ResendMessage(val eventId: String) : RoomDetailActions()
     data class RemoveFailedEcho(val eventId: String) : RoomDetailActions()
+
+    data class ReportContent(val eventId: String, val reason: String, val spam: Boolean = false, val inappropriate: Boolean = false) : RoomDetailActions()
+
     object ClearSendQueue : RoomDetailActions()
     object ResendAll : RoomDetailActions()
 }