+++ CHANGES.md
@@ -9,6 +9,7 @@ Improvements:
  - Do not upload file too big for the homeserver (#587)
  - Handle read markers (#84)
  - Mark all messages as read (#396)
+ - Add ability to report content (#515)
 
 Other changes:
  - Accessibility improvements to read receipts in the room timeline and reactions emoji chooser